Refactor/folder structure revision&bug fixes

+++ app/api/playlist/[id]/like/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { PlaylistLikeType } from "@/libs/types/common/Song&PlaylistType";
+import { PlaylistLikeType } from "@/libs/types/song&playlistType";
 import { prisma } from "@/prisma/client";
 
 export async function POST(req: Request) {
@@ -7,11 +7,17 @@ export async function POST(req: Request) {
     const request: PlaylistLikeType = await req.json().then((data) => data);
     const { userId, playlistId } = request;
 
+    if (!userId || !playlistId)
+      return new NextResponse(JSON.stringify({ message: "fail" }), {
+        status: 400,
+        statusText: "Bad Request",
+      });
+
     const existingLike = await prisma.playlistLikedByUsers.findFirst({
       where: { userId, playlistId },
     });
 
-    if (existingLike) {
+    if (existingLike && userId) {
       await prisma.playlistLikedByUsers.delete({
         where: { userId_playlistId: { userId, playlistId } },
       });
@@ -25,9 +31,27 @@ export async function POST(req: Request) {
         },
       });
 
+      const likedBy = await prisma.playlist.findUnique({
+        where: { id: playlistId },
+        select: {
+          likedBy: {
+            select: {
+              userId: true,
+              playlistId: true,
+              user: {
+                select: { nickname: true, profilePic: true },
+              },
+            },
+          },
+        },
+      });
+
       return new NextResponse(
         JSON.stringify({
           message: "unlike success",
+          userId,
+          playlistId,
+          likedBy: likedBy?.likedBy || [],
         }),
         { status: 200, statusText: "OK" },
       );
@@ -39,6 +63,7 @@ export async function POST(req: Request) {
         playlistId,
       },
     });
+
     await prisma.playlist.update({
       where: { id: playlistId },
       data: {
@@ -48,9 +73,27 @@ export async function POST(req: Request) {
       },
     });
 
+    const likedBy = await prisma.playlist.findUnique({
+      where: { id: playlistId },
+      select: {
+        likedBy: {
+          select: {
+            userId: true,
+            playlistId: true,
+            user: {
+              select: { nickname: true, profilePic: true },
+            },
+          },
+        },
+      },
+    });
+
     return new NextResponse(
       JSON.stringify({
         message: "like success",
+        userId,
+        playlistId,
+        likedBy: likedBy?.likedBy || [],
       }),
       { status: 200, statusText: "OK" },
     );
